Fix alignment mask calculations and remove int32_t cast
- Changed alignedEnd calculation from ~(nuint)16 to ~(nuint)(16 - 1) for clarity
- Both alignment masks now use consistent (16 - 1) pattern
- Removed (int32_t) cast from ASSERT in allocheap.cpp to fix gcc build break
- ASSERT now directly compares alignment <= OS_PAGE_SIZE without cast
- Both are size_t types so no cast needed

+++ src/coreclr/nativeaot/Runtime/allocheap.cpp
@@ -88,7 +88,7 @@ uint8_t * AllocHeap::_Alloc(
     )
 {
     ASSERT((alignment & (alignment - 1)) == 0); // Power of 2 only.
-    ASSERT((int32_t)alignment <= OS_PAGE_SIZE);          // Can't handle this right now.
+    ASSERT(alignment <= OS_PAGE_SIZE);          // Can't handle this right now.
 
     CrstHolder lock(&m_lock);
